// contentful.d.ts
// THIS FILE IS AUTOMATICALLY GENERATED. DO NOT MODIFY IT.

import { Asset, Entry } from 'contentful';
import { Document } from '@contentful/rich-text-types';

export interface IBooksFields {
  /** poster */
  poster?: Asset | undefined;

  /** booktitle */
  booktitle?: string | undefined;

  /** Show description */
  showDescription?: string | undefined;

  /** description */
  description?: string | undefined;

  /** button */
  button?: string | undefined;
}

export interface IBooks extends Entry<IBooksFields> {
  sys: {
    id: string;
    type: string;
    createdAt: string;
    updatedAt: string;
    locale: string;
    contentType: {
      sys: {
        id: 'books';
        linkType: 'ContentType';
        type: 'Link';
      };
    };
  };
}

export interface IMainFields {
  /** title */
  title?: string | undefined;

  /** description */
  description: Document;

  /** background */
  background?: Asset | undefined;

  /** Footer */
  footer?: string | undefined;
}

export interface IMain extends Entry<IMainFields> {
  sys: {
    id: string;
    type: string;
    createdAt: string;
    updatedAt: string;
    locale: string;
    contentType: {
      sys: {
        id: 'main';
        linkType: 'ContentType';
        type: 'Link';
      };
    };
  };
}

export type CONTENT_TYPE = 'books' | 'main';

export type LOCALE_CODE = 'en-US';

export type CONTENTFUL_DEFAULT_LOCALE_CODE = 'en-US';